+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- code: refactor to use otel tracing instead of prometheus (removed mc-metrics, added mc-analytics)
 - fix(version constants): 0.13.2 was mapped to wrong constants
 - fix(compilation): devnet contract artifacts are not compiled by `cargo build` anymore
 - feat: add fgw get_block_traces